requirements.txt
Kivy==2.3.1
Kivy-Garden==0.1.5
kivymd==1.2.0
firebase-admin==6.8.0
google-api-python-client==2.169.0
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
fpdf==1.7.2
reportlab==4.4.0
requests==2.32.3
python-dotenv
Pillow==11.2.1
